Create splitByAllowance middleware that splits a list of IDs into allowed and disallowed by checking user permissions

+++ backend/src/modules/organizations/index.ts
@@ -455,70 +455,13 @@ const organizationsRoutes = app
    * Delete organizations
    */
   .openapi(deleteOrganizationsRouteConfig, async (ctx) => {
-    const { ids } = ctx.req.valid('query');
-    const user = ctx.get('user');
+    // * Extract allowed and disallowed ids
+    const allowedIds = ctx.get('allowedIds');
+    const disallowedIds = ctx.get('disallowedIds')
 
-    // * Convert the ids to an array
-    const organizationIds = Array.isArray(ids) ? ids : [ids];
-
-    const errors: ErrorType[] = [];
-
-    // * Get the organizations and the user role
-    const targets = await db
-      .select({
-        organization: organizationsTable,
-        userRole: membershipsTable.role,
-      })
-      .from(organizationsTable)
-      .leftJoin(membershipsTable, and(eq(membershipsTable.organizationId, organizationsTable.id), eq(membershipsTable.userId, user.id)))
-      .where(inArray(organizationsTable.id, organizationIds));
-
-    // * Check if the organizations exist
-    for (const id of organizationIds) {
-      if (!targets.some((target) => target.organization.id === id)) {
-        errors.push(
-          createError(ctx, 404, 'not_found', 'warn', 'ORGANIZATION', {
-            organization: id,
-          }),
-        );
-      }
-    }
-
-    // * Filter out organizations that the user doesn't have permission to delete
-    const allowedTargets = targets.filter((target) => {
-      const organizationId = target.organization.id;
-
-      if (user.role !== 'ADMIN' && target.userRole !== 'ADMIN') {
-        errors.push(
-          createError(ctx, 403, 'delete_forbidden', 'warn', 'ORGANIZATION', {
-            organization: organizationId,
-          }),
-        );
-        return false;
-      }
-
-      return true;
-    });
-
-    // * If the user doesn't have permission to delete any of the organizations, return an error
-    if (allowedTargets.length === 0) {
-      return ctx.json(
-        {
-          success: false,
-          errors: errors,
-        },
-        200,
-      );
-    }
-
-    // // * Deleted organizations members
-    // await db.delete(membershipsTable).where(
-    //   inArray(
-    //     organizationsTable.id,
-    //     allowedTargets.map((target) => target.organization.id),
-    //   ),
-    // );
-
+    // * Map errors of workspaces user is not allowed to delete 
+    const errors: ErrorType[] = disallowedIds.map(id => createError(ctx, 404, 'not_found', 'warn', 'ORGANIZATION', { organization: id }));
+    
     // * Get members
     const organizationsMembers = await db
       .select({ id: membershipsTable.userId })
@@ -526,31 +469,22 @@ const organizationsRoutes = app
       .where(
         and(
           eq(membershipsTable.type, 'ORGANIZATION'),
-          inArray(
-            membershipsTable.organizationId,
-            allowedTargets.map((target) => target.organization.id),
-          ),
+          inArray(membershipsTable.organizationId, allowedIds),
         ),
       );
 
     // * Delete the organizations
-    await db.delete(organizationsTable).where(
-      inArray(
-        organizationsTable.id,
-        allowedTargets.map((target) => target.organization.id),
-      ),
-    );
+    await db.delete(organizationsTable).where(inArray(organizationsTable.id, allowedIds));
 
     // * Send SSE events for the organizations that were deleted
-    for (const { organization } of allowedTargets) {
+    for (const id of allowedIds) {
       // * Send the event to the user if they are a member of the organization
-
       if (organizationsMembers.length > 0) {
         const membersId = organizationsMembers.map((member) => member.id).filter(Boolean) as string[];
-        sendSSEToUsers(membersId, 'remove_entity', organization);
+        sendSSEToUsers(membersId, 'remove_entity', { id, type: 'ORGANIZATION' });
       }
 
-      logEvent('Organization deleted', { organization: organization.id });
+      logEvent('Organization deleted', { organization: id });
     }
 
     return ctx.json(
